Moves new circular percentiles class into circular_kernel module.
