+++ lib/improver/nbhood/percentiles.py
@@ -36,7 +36,7 @@ import numpy as np
 
 from improver.utilities.cube_manipulation import concatenate_cubes
 from improver.nbhood.nbhood import Utilities as NBUtilities
-from improver.nbhood.circular_kernel import Utilities
+from improver.nbhood.circular_kernel import (Utilities, CircularKernelNumpy)
 from improver.utilities.spatial import (
     convert_distance_into_number_of_grid_cells)
 from improver.percentile import PercentileConverter
@@ -222,162 +222,3 @@ class NeighbourhoodPercentiles(object):
                                              coords_to_slice_over=["time"])
 
         return cube_new
-
-
-class CircularKernelNumpy(object):
-    """
-    Methods for use in calculating percentiles from a 2D circular
-    neighbourhood.
-
-    A maximum kernel radius of 500 grid cells is imposed in order to
-    avoid computational ineffiency and possible memory errors.
-    """
-    def __init__(self,
-                 percentiles=PercentileConverter.DEFAULT_PERCENTILES):
-        """
-
-        Parameters
-        ----------
-
-        percentiles : list (optional)
-            Percentile values at which to calculate; if not provided uses
-            DEFAULT_PERCENTILES from percentile module.
-        """
-        self.percentiles = percentiles
-
-    def __repr__(self):
-        """Represent the configured class instance as a string."""
-        result = ('<CircularKernelNumpy: percentiles: {}>')
-        return result.format(self.percentiles)
-
-    def run(self, cube, ranges):
-        """
-        Method to apply a circular kernel to the data within the input cube in
-        order to derive percentiles over the kernel.
-
-        Parameters
-        ----------
-        cube : Iris.cube.Cube
-            Cube containing array to apply processing to.
-        ranges : Int
-            Number of grid cells in the x and y direction used to create
-            the kernel.
-
-        Returns
-        -------
-        outcube : Iris.cube.Cube
-            Cube containing the percentile fields.
-            Has percentile as an added dimension.
-
-        """
-        # Take data array and identify X and Y axes indices
-        try:
-            for coord_name in ['projection_x_coordinate',
-                               'projection_y_coordinate']:
-                cube.coord(coord_name)
-        except CoordinateNotFoundError:
-            raise ValueError("Invalid grid: projection_x/y coords required")
-        ranges = int(ranges)
-        if ranges < 1:
-            raise ValueError("Range size too small. {} < 1".format(ranges))
-        ranges_xy = np.array([ranges]*2)
-        ranges_tuple = tuple([ranges]*2)
-        kernel = Utilities.circular_kernel(ranges_xy, ranges_tuple,
-                                           weighted_mode=False)
-
-        # Loop over each 2D slice to reduce memory demand and derive
-        # percentiles on the kernel. Will return an extra dimension.
-        pctcubelist = iris.cube.CubeList()
-        for slice_2d in cube.slices(['projection_y_coordinate',
-                                     'projection_x_coordinate']):
-            # Create a 1D data array padded with repeats of the local boundary
-            # mean.
-            padded = np.pad(slice_2d.data, ranges, mode='mean',
-                            stat_length=ranges)
-            padshape = np.shape(padded)  # Store size to make unflatten easier
-            padded = padded.flatten()
-            # Add 2nd dimension with each point's neighbourhood points along it
-            nbhood_slices = [
-                np.roll(padded, padshape[1]*j+i)
-                for i in range(-ranges, ranges+1)
-                for j in range(-ranges, ranges+1)
-                if kernel[..., i+ranges, j+ranges] > 0.]
-            # Collapse this dimension into percentiles (a new 2nd dimension)
-            perc_data = np.percentile(nbhood_slices, self.percentiles, axis=0)
-            # Return to 3D
-            perc_data = perc_data.reshape(
-                len(self.percentiles), padshape[0], padshape[1])
-            # Create a cube for these data:
-            pctcube = self.make_percentile_cube(slice_2d)
-            # And put in data, removing the padding
-            pctcube.data = perc_data[:, ranges:-ranges, ranges:-ranges]
-            pctcubelist.append(pctcube)
-        result = pctcubelist.merge_cube()
-        result = self.check_coords(result, cube)
-        return result
-
-    @staticmethod
-    def check_coords(cube, cube_orig):
-        """Checks the coordinates of cube match those of cube_orig
-        and promotes any that are not dimensions.
-        This function expects that cube will have an additional
-        "percentiles" dimension.
-
-        Parameters
-        ----------
-        cube : Iris.cube.Cube
-            Cube to ensure compliance in. May be modified if not compliant.
-
-        cube_orig : Iris.cube.Cube
-            Cube to ensure compliance against. Will NOT be modified.
-
-        Returns
-        -------
-        cube : Iris.cube.Cube
-            Cube after ensuring compliance.
-
-        Exceptions
-        -------
-        Raises ValueError if cube cannot be made compliant.
-        """
-
-        # Promote any missing dimension coords from auxilliary coords
-        for coord in cube_orig.coords():
-            if len(cube_orig.coord_dims(coord)) == 0:
-                continue
-            try:
-                cube.coord_dims(coord)[0]
-            except IndexError:
-                cube = iris.util.new_axis(cube, coord)
-        # Now check axis order
-        required_order = list(np.shape(cube.data))
-        for indx, coord in enumerate(cube_orig.coords()):
-            if len(cube_orig.coord_dims(coord)) == 0:
-                continue
-            required_order[indx+1] = cube.coord_dims(coord)[0]
-        required_order[0] = cube.coord_dims("percentiles")[0]
-        cube.transpose(required_order)
-        return cube
-
-    def make_percentile_cube(self, cube):
-        """Returns a cube with the same metadata as the sample cube
-        but with an added percentile dimension.
-
-        Parameters
-        ----------
-        cube : Iris.cube.Cube
-            Cube to copy meta data from.
-
-        Returns
-        -------
-        cube : Iris.cube.Cube
-            Cube like input but with added percentiles coordinate.
-            Each slice along this coordinate is identical.
-        """
-        pctcubelist = iris.cube.CubeList()
-        for pct in self.percentiles:
-            pctcube = cube.copy()
-            pctcube.add_aux_coord(iris.coords.DimCoord(
-                pct, long_name='percentiles', units='%'))
-            pctcubelist.append(pctcube)
-        return pctcubelist.merge_cube()
